- Moved ITopOrderService and TopOrderServiceStub out of the redundant"service" package and into the existing "services" package. - Deleted the now-empty "service" package - Renamed the autowired service in the controller to match the name of the actual class
