+++ src/main/java/com/example/icecreamup/IcecreamupController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.example.icecreamup.dto.ProductDTO;
-import com.example.icecreamup.service.ITopOrderService;
+import com.example.icecreamup.services.ITopOrderService;
 
 import org.springframework.web.bind.annotation.PostMapping;
 
@@ -23,7 +23,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 public class IcecreamupController {
 	
 	@Autowired
-	private ITopOrderService topOrderService;
+	private ITopOrderService topOrderServiceStub;
 	
 	@RequestMapping(value="/start", method=RequestMethod.GET)
 	public String read() {
@@ -37,10 +37,11 @@ public class IcecreamupController {
 
 	@RequestMapping(value="/topOrders", method=RequestMethod.GET)
 	public String topOrder(Model model) {
-		ProductDTO productDTO = topOrderService.fetchById(50);
+		ProductDTO productDTO = topOrderServiceStub.fetchById(50);
 		model.addAttribute("productDTO", productDTO);
 		return "topOrders";
 	}
+	
 	@RequestMapping("/")
 	public String index() {
