+++ src/main/java/com/example/icecreamup/services/ITopOrderService.java
@@ -1,0 +1,16 @@
+package com.example.icecreamup.services;
+
+import com.example.icecreamup.dto.ProductDTO;
+/* This is the Top Order interface .
+ * 
+ * 
+ */
+public interface ITopOrderService {
+
+	/**Get Flavors from persistence layer.
+	 * @param id an unique lookup
+	 * @return a flavor with a matching id.
+	 */
+	ProductDTO fetchById(int id);
+}
+
